organized createprofile components
